+++ src/main/java/lotto/domain/AnswerLotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.message.ExceptionMessage;
