+++ src/main/java/lotto/domain/LottoPurchaseInfo.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
