+++ src/main/java/lotto/message/ExceptionMessage.java
@@ -1,6 +1,6 @@
 package lotto.message;
 
-import lotto.LottoOption;
+import lotto.domain.LottoOption;
 
 public class ExceptionMessage {
     public static final String INVALID_SIZE = "허용되는 로또 숫자 개수는 " + LottoOption.LOTTO_SIZE + "입니다.";
