+++ src/test/java/lotto/AnswerLottoTest.java
@@ -2,6 +2,9 @@ package lotto;
 
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.domain.AnswerLotto;
+import lotto.domain.BonusNumber;
+import lotto.domain.Lotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
