+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import java.util.stream.Stream;
+import lotto.domain.Lotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
