+++ src/test/java/lotto/WalletTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Wallet;
 import lotto.message.ExceptionMessage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
